Linksphere_Application
To initiate the project, open a new terminal in VS Code and enter "npm start". Next, set up an Expo router to navigate the application's content to the desired hosting platforms, such as Android, iOS, or web. Then, configure the Apollo client and neonDB to integrate with the GraphQL API. After these steps, the application will be ready for use.